[RUBY-3154] refactoring code to rid of rspec warning
