+++ spec/models/waste_exemptions_engine/concerns/can_convert_pence_to_pounds.rb
